+++ jdroid-android/src/main/java/com/jdroid/android/firebase/dynamiclink/DynamicLinkBuilder.java
@@ -15,7 +15,7 @@ public class DynamicLinkBuilder {
 
 	// Optional: The versionCode of the minimum version of your app that can open the link.
 	// If the installed app is an older version, the user is taken to the Play Store to upgrade the app.
-	private Long minVersionCode;
+	private Integer minVersionCode;
 
 	// Optional: The link to open when the app isn't installed. Specify this to do something other than install your app from the Play Store when the app isn't installed,
 	// such as open the mobile web version of the content, or display a promotional page for your app.
